Refactoring tpu_api_initializer
