+++ tensorflow/core/tpu/BUILD
@@ -156,6 +156,16 @@ cc_library(
     srcs = ["tpu_initializer_helper.cc"],
     hdrs = ["tpu_initializer_helper.h"],
     deps = [
+	":libtftpu_header",
+        ":tpu_api",
+        ":tpu_api_dlsym_set_fn",
+        ":tpu_executor_init_fns",
+        ":tpu_library_init_fns",
+        ":tpu_ops_c_api_hdrs",
+        "//tensorflow/core:lib",
+        "//tensorflow/core/platform/cloud:gcs_file_system",
+        "//tensorflow/stream_executor/tpu:tpu_executor",
+        "//tensorflow/stream_executor/tpu:tpu_executor_c_api_hdrs",
         "//tensorflow/core/platform:logging",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/synchronization",
